requirements.txt
librosa
numpy
scipy
soundfile
pydub
matplotlib
opencv-python
torch
torchvision
pillow